refactor: 부족한 부분 채웠습니다 #30
