+++ src/main/java/com/example/momogum/service/userProfileService/UserProfileServiceImpl.java
@@ -1,4 +1,4 @@
-package com.example.momogum.service;
+package com.example.momogum.service.userProfileService;
 
 
 import com.example.momogum.apiPayLoad.code.status.ErrorStatus;
@@ -8,21 +8,20 @@ import com.example.momogum.repository.followRepo.FollowRepository;
 import com.example.momogum.repository.userEntityRepo.UserEntityRepository;
 import com.example.momogum.web.dto.FollowDTO;
 import com.example.momogum.web.dto.user.UserDTO;
+import org.springframework.transaction.annotation.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 
 @Service
 @RequiredArgsConstructor
-public class UserProfileService {
+public class UserProfileServiceImpl implements UserProfileService {
 
   private final UserEntityRepository userEntityRepository;
   private final FollowRepository followRepository;
 
-  /**
-   * 유저 정보 ID로 조회
-   **/
-
+  @Override
+  @Transactional(readOnly = true)
   public UserDTO.UserResponseDTO getUserProfile(Long userId) {
     UserEntity user = userEntityRepository.findById(userId)
         .orElseThrow(() -> new GeneralException(ErrorStatus._BAD_REQUEST));
@@ -38,11 +37,8 @@ public class UserProfileService {
         .build();
   }
 
-  /**
-   * 팔로워/팔로잉 수 조회
-   * 0. 팔로워/팔로잉 수 카운트
-   */
-
+  @Override
+  @Transactional(readOnly = true)
   public FollowDTO.FollowStatsDTO getFollowStats(Long userId) {
     UserEntity user = userEntityRepository.findById(userId)
         .orElseThrow(() -> new GeneralException(ErrorStatus._BAD_REQUEST));
@@ -56,26 +52,12 @@ public class UserProfileService {
         .build();
   }
 
-  /**
-   * 팔로워/ 팔로잉 멤버 조회
-   * 0. 멤버 조회
-   * 1. 멤버 프로필
-   */
-
-  /**
-   * 유저 프로필 수정
-   * 0. 프로필 이미지 편집
-   * 1. 기본 이미지
-   * 2. 갤러리에서 사진 첨부
-   * 3. 이름, 닉네임, 한줄 소개 수정
-   */
-
+  @Override
+  @Transactional
   public UserDTO.UserEditDTO updateUserProfile(Long userId, UserDTO.UserEditDTO request) {
-    // 유저 데이터 조회
     UserEntity user = userEntityRepository.findById(userId)
         .orElseThrow(() -> new GeneralException(ErrorStatus._BAD_REQUEST));
 
-    // 변경사항 존재할 경우 업데이트
     if (request.getNickname() != null && !request.getNickname().equals(user.getNickname())) {
       user.setNickname(request.getNickname());
     }
@@ -87,7 +69,6 @@ public class UserProfileService {
         user.setAbout(request.getAbout());
       }
     } else {
-      // 한 줄 소개의 경우 값이 없으면 기본값(빈 문자열)으로 설정
       user.setAbout("");
     }
 
@@ -97,7 +78,4 @@ public class UserProfileService {
         .about(user.getAbout() != null ? user.getAbout() : "")
         .build();
   }
-
-
-
 }
